+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/UidTest.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Tests\Functional;
 
+use Symfony\Bundle\FrameworkBundle\Test\ExceptionSubscriber;
 use Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\Controller\UidController;
 use Symfony\Component\Uid\Ulid;
 use Symfony\Component\Uid\UuidV1;
@@ -31,12 +32,16 @@ class UidTest extends AbstractWebTestCase
 
     public function testArgumentValueResolverDisabled()
     {
-        $this->expectException(\TypeError::class);
-        $this->expectExceptionMessage('Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\Controller\UidController::anyFormat(): Argument #1 ($userId) must be of type Symfony\Component\Uid\UuidV1, string given');
-
         $client = $this->createClient(['test_case' => 'Uid', 'root_config' => 'config_disabled.yml']);
 
         $client->request('GET', '/1/uuid-v1/'.new UuidV1());
+        $this->assertSame(500, $client->getResponse()->getStatusCode());
+        $exceptions = ExceptionSubscriber::shiftAll();
+        $this->assertCount(1, $exceptions);
+        $exception = reset($exceptions);
+
+        $this->assertInstanceOf(\TypeError::class, $exception);
+        $this->assertStringContainsString('Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\Controller\UidController::anyFormat(): Argument #1 ($userId) must be of type Symfony\Component\Uid\UuidV1, string given', $exception->getMessage());
     }
 
     public function testArgumentValueResolverEnabled()