+++ src/Symfony/Bundle/FrameworkBundle/Tests/Kernel/MicroKernelTraitTest.php
@@ -122,6 +122,7 @@ class MicroKernelTraitTest extends TestCase
                 $c->extension('framework', [
                     'http_method_override' => false,
                     'router' => ['utf8' => true],
+                    'catch_all_throwables' => true,
                 ]);
                 $c->services()->set('logger', NullLogger::class);
             }